Patch fix clippy 35 warnings and errors (#620)
* clippy patch 240318 afternoon1 compilable runnable

* patch-fix-clippy-240318afternoon2

* fmt

---------

+++ kernel/src/arch/x86_64/interrupt/ipi.rs
@@ -102,7 +102,7 @@ impl ArchIpiTarget {
     #[inline(always)]
     fn cpu_id_to_apic_id(cpu_id: ProcessorId) -> x86::apic::ApicId {
         if CurrentApic.x2apic_enabled() {
-            x86::apic::ApicId::X2Apic(cpu_id.data() as u32)
+            x86::apic::ApicId::X2Apic(cpu_id.data())
         } else {
             x86::apic::ApicId::XApic(cpu_id.data() as u8)
         }
